phase-yb: fix y7 delivery hardening gaps
