+++ docs/phase-Yb/sprint-Y7.md
@@ -1,7 +1,7 @@
 ---
 id: Y.7
 title: Degraded Delivery Contract Hardening
-status: planned
+status: complete
 branch: feature/pYb-s7-degraded-delivery-contract-hardening
 worktree: ../atm-core-worktrees/feature/pYb-s7-degraded-delivery-contract-hardening
 target: integrate/phase-Yb
@@ -93,22 +93,22 @@ policy. These are listed here as context for Y.7 developers. Only item 2 is a
 Y.7 removal obligation; items 1, 3, and 4 are addressed in later sprints as
 assigned in the removal ledger.
 
-1. [delivery_policy.rs](</Users/randlee/Documents/github/atm-core-worktrees/integrate/phase-Y/crates/atm-core/src/delivery_policy.rs:61>)
+1. `crates/atm-core/src/delivery_policy.rs:61`
    - `DeliveryRecipientSnapshot::fallback_claude`
    - ledger: `YB-RM-014`, sprint: `Y.9`
    - inconsistency: the name and behavior silently default an unresolved route
      to `DeliveryHarnessPath::ClaudeCode`
-2. [send/persistence.rs](</Users/randlee/Documents/github/atm-core-worktrees/integrate/phase-Y/crates/atm-core/src/send/persistence.rs:85>)
+2. `crates/atm-core/src/send/persistence.rs:85`
    - `recipient.allows_claude_jsonl_append()` branch inside
      `recover_after_sqlite_failure(...)`
    - ledger: `YB-RM-002` / `YB-RM-003`, sprint: `Y.7`
    - inconsistency: persistence still decides a harness-specific outward path
-3. [service_runtime.rs](</Users/randlee/Documents/github/atm-core-worktrees/integrate/phase-Y/crates/atm-core/src/service_runtime.rs:181>)
+3. `crates/atm-core/src/service_runtime.rs:181`
    - early return in `refresh_compat_inbox_projection(...)`
    - ledger: `YB-RM-017`, sprint: `Y.10`
    - inconsistency: the Claude writer is still asked to inspect a non-Claude
      request and no-op
-4. [service_runtime.rs](</Users/randlee/Documents/github/atm-core-worktrees/integrate/phase-Y/crates/atm-core/src/service_runtime.rs:202>)
+4. `crates/atm-core/src/service_runtime.rs:202`
    - early return in `append_compat_inbox_message(...)`
    - ledger: `YB-RM-019`, sprint: `Y.9`
    - inconsistency: the low-level Claude append seam still receives
@@ -223,3 +223,12 @@ cargo build --workspace
 cargo test --workspace
 git diff --check
 ```
+
+## Validation Record
+
+- branch closeout validated with:
+  - `cargo fmt --all --check`
+  - `python3 .just/run_lint.py all`
+  - `cargo build --workspace`
+  - `cargo test --workspace`
+  - `git diff --check`
